refactor: Split user profile hooks
