+++ src/App.tsx
@@ -3,8 +3,9 @@ import { Router, Redirect, Route, Switch } from 'react-router-dom'
 import { ResetCSS } from '@pancakeswap/uikit'
 import BigNumber from 'bignumber.js'
 import useEagerConnect from 'hooks/useEagerConnect'
-import { useFetchProfile, usePollBlockNumber } from 'state/hooks'
+import { usePollBlockNumber } from 'state/hooks'
 import { usePollCoreFarmData } from 'state/farms/hooks'
+import { useFetchProfile } from 'state/profile/hooks'
 import { DatePickerPortal } from 'components/DatePicker'
 import GlobalStyle from './style/Global'
 import Menu from './components/Menu'
