+++ src/state/achievements/hooks.ts
@@ -1,0 +1,22 @@
+import { useEffect } from 'react'
+import { useWeb3React } from '@web3-react/core'
+import { useSelector } from 'react-redux'
+import { useAppDispatch } from 'state'
+import { State, AchievementState } from '../types'
+import { fetchAchievements } from '.'
+
+export const useFetchAchievements = () => {
+  const { account } = useWeb3React()
+  const dispatch = useAppDispatch()
+
+  useEffect(() => {
+    if (account) {
+      dispatch(fetchAchievements(account))
+    }
+  }, [account, dispatch])
+}
+
+export const useAchievements = () => {
+  const achievements: AchievementState['data'] = useSelector((state: State) => state.achievements.data)
+  return achievements
+}
