+++ src/state/teams/hooks.ts
@@ -1,0 +1,28 @@
+import { useEffect } from 'react'
+import { useSelector } from 'react-redux'
+import { useAppDispatch } from 'state'
+import { Team } from 'config/constants/types'
+import { State, TeamsState } from '../types'
+import { fetchTeam, fetchTeams } from '.'
+
+export const useTeam = (id: number) => {
+  const team: Team = useSelector((state: State) => state.teams.data[id])
+  const dispatch = useAppDispatch()
+
+  useEffect(() => {
+    dispatch(fetchTeam(id))
+  }, [id, dispatch])
+
+  return team
+}
+
+export const useTeams = () => {
+  const { isInitialized, isLoading, data }: TeamsState = useSelector((state: State) => state.teams)
+  const dispatch = useAppDispatch()
+
+  useEffect(() => {
+    dispatch(fetchTeams())
+  }, [dispatch])
+
+  return { teams: data, isInitialized, isLoading }
+}
