+++ src/views/Collectibles/components/GlobalCheckClaimStatus.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef, useState } from 'react'
 import { useLocation } from 'react-router-dom'
 import { useWeb3React } from '@web3-react/core'
 import { useModal } from '@pancakeswap/uikit'
-import { useProfile } from 'state/hooks'
+import { useProfile } from 'state/profile/hooks'
 import { useEasterNftContract } from 'hooks/useContract'
 import NftGiveawayModal from './NftGiveawayModal'
