+++ src/views/Collectibles/components/NftGiveawayModal.tsx
@@ -7,7 +7,7 @@ import { delay } from 'lodash'
 import { useTranslation } from 'contexts/Localization'
 import nftList from 'config/constants/nfts'
 import { Nft } from 'config/constants/types'
-import { useProfile } from 'state/hooks'
+import { useProfile } from 'state/profile/hooks'
 import { Profile } from 'state/types'
 import { teamNftMap } from './NftCard/EasterNftCard'
