+++ src/views/Ifos/components/IfoFoldableCard/IfoPoolCard/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useTranslation } from 'contexts/Localization'
 import { Card, CardBody, CardHeader, Text, useTooltip, HelpIcon, Flex } from '@pancakeswap/uikit'
 import { Ifo, PoolIds } from 'config/constants/types'
-import { useProfile } from 'state/hooks'
+import { useProfile } from 'state/profile/hooks'
 import { PublicIfoData, WalletIfoData } from 'views/Ifos/types'
 import { EnableStatus } from '../types'
 import IfoCardTokens from './IfoCardTokens'
