+++ src/views/Profile/components/AchievementsList.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { Flex, Heading } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
-import { useAchievements } from 'state/hooks'
+import { useAchievements } from 'state/achievements/hooks'
 import AchievementCard from './AchievementCard'
 
 const Grid = styled.div`
