+++ src/views/Profile/components/ClaimPointsCallout .tsx
@@ -3,7 +3,7 @@ import { sumBy } from 'lodash'
 import { useAppDispatch } from 'state'
 import { useWeb3React } from '@web3-react/core'
 import { Card, CardBody, CardHeader, Flex, Heading, PrizeIcon } from '@pancakeswap/uikit'
-import { useProfile } from 'state/hooks'
+import { useProfile } from 'state/profile/hooks'
 import { Achievement } from 'state/types'
 import { addPoints } from 'state/profile'
 import { addAchievement } from 'state/achievements'
