+++ src/views/Profile/index.tsx
@@ -3,7 +3,8 @@ import { Route } from 'react-router-dom'
 import { useWeb3React } from '@web3-react/core'
 import Page from 'components/Layout/Page'
 import PageLoader from 'components/Loader/PageLoader'
-import { useFetchAchievements, useProfile } from 'state/hooks'
+import { useProfile } from 'state/profile/hooks'
+import { useFetchAchievements } from 'state/achievements/hooks'
 import ProfileCreation from './ProfileCreation'
 import Header from './components/Header'
 import TaskCenter from './TaskCenter'
