+++ src/views/Teams/components/TeamHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Heading, Text } from '@pancakeswap/uikit'
-import { useProfile } from 'state/hooks'
+import { useProfile } from 'state/profile/hooks'
 import { useTranslation } from 'contexts/Localization'
 import HeaderWrapper from 'views/Profile/components/HeaderWrapper'
 import NoProfileCard from './NoProfileCard'
