+++ src/views/Teams/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { AutoRenewIcon, Flex, Heading } from '@pancakeswap/uikit'
 import orderBy from 'lodash/orderBy'
-import { useTeams } from 'state/hooks'
+import { useTeams } from 'state/teams/hooks'
 import Page from 'components/Layout/Page'
 import { useTranslation } from 'contexts/Localization'
 import TeamListCard from './components/TeamListCard'
